- Separation of User management service into Staff + Unit + Role services. - Removing redundancies form previous commits

+++ backend/src/main/java/com/Folio/attendance_app/service/SessionService.java
@@ -20,7 +20,7 @@ public class SessionService {
     }
 
     public boolean isAdmin(HttpSession session) {
-    Integer staffId = (Integer) session.getAttribute(SessionConstants.STAFF_ID);
+    Long staffId = (Long) session.getAttribute(SessionConstants.STAFF_ID);
     if (staffId == null) return false;
     Staff staff = userManagementService.getStaffById(staffId);
     return staff != null && staff.isAdmin();
